create useLegacySelectedPosts() hook to keep post-picker-control lean
